Add files via upload
The code in the two jupyter notebooks are for MLP models created for Regression Problem for a particular feature representation technique. We have created multiple jupyter notebooks containing MLP models for both Regression and Classification Problems corresponding to different feature representation techniques which have a similar code but with some changes in parameter values.